Add files via upload
The purpose of website is to find amount of heat the substance is producing with the help of various scenarios like time, electricity consumption etc and making user aware that how much amount of heat his actions are giving to the environment. Also if any area is found to be more hazardous then giving information to the nearest environmentalist present over there.